Create user api on front end
