+++ app/src/components/authentication/ProtectedRoute.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import {Navigate, useLocation} from 'react-router-dom';
 import {useAuth} from '../../app/hooks/useAuth';
 import {Loading} from '../common';
-import {useUserQuery} from '../../services/auth';
+import {useUserQuery} from '../../services/user';
 
 export const ProtectedRoute = ({children}: {children: JSX.Element}) => {
   const {user} = useAuth();
